[storage-file-share] typescript samples in shape
